tweak rpc/readme.md
